refacto(prb): move to prb package
