+++ pkg/prb/probe.go
@@ -1,4 +1,4 @@
-package svr
+package prb
 
 import (
 	"context"
@@ -13,7 +13,7 @@ type Probe struct {
 	period  time.Duration
 	ticker  *time.Ticker
 	checker chk.Checker
-	ch      chan *ProbeTick
+	ch      chan *Tick
 }
 
 // NewProbe with period and checker.
@@ -22,8 +22,8 @@ func NewProbe(name string, period time.Duration, checker chk.Checker) *Probe {
 }
 
 // Start the probe.
-func (p *Probe) Start(ctx context.Context) chan *ProbeTick {
-	p.ch = make(chan *ProbeTick)
+func (p *Probe) Start(ctx context.Context) chan *Tick {
+	p.ch = make(chan *Tick)
 	p.ticker = time.NewTicker(p.period)
 
 	go p.start(ctx)
@@ -40,7 +40,7 @@ func (p *Probe) start(ctx context.Context) {
 		case <-ctx.Done():
 			return
 		case <-p.ticker.C:
-			p.ch <- NewProbeTick(p.name, p.checker.Check(ctx))
+			p.ch <- NewTick(p.name, p.checker.Check(ctx))
 		}
 	}
 }
